refactor(tests): separate daemon integration
